remove corpsum page -> everything in home
